@@ -9,7 +9,8 @@ import { HttpClientModule } from '@angular/common/http';
 import { ToastrModule } from 'ngx-toastr';
 import { BrowserAnimationsModule } from '@angular/platform-browser/animations';
 import { ClipboardModule } from 'ngx-clipboard';
-
+import { NgbModule, NgbCarouselModule, NgbSlide } from '@ng-bootstrap/ng-bootstrap';
+import { EventDetailComponent } from './module/front/event-detail/event-detail.component';
 @NgModule({
   declarations: [
     AppComponent,
@@ -22,7 +23,9 @@ import { ClipboardModule } from 'ngx-clipboard';
     HttpClientModule,
     ToastrModule.forRoot(),
     BrowserAnimationsModule,
-    ClipboardModule
+    ClipboardModule,
+    NgbModule,
+    NgbCarouselModule,
   ],
   providers: [],
   bootstrap: [AppComponent]
