@@ -3,7 +3,7 @@ import { CommonModule } from '@angular/common';
 
 import { PdfNotFoundRoutingModule } from './pdf-not-found-routing.module';
 import { PdfNotFoundComponent } from './pdf-not-found.component';
-
+import { NgbModule, NgbCarouselModule, NgbSlide } from '@ng-bootstrap/ng-bootstrap';
 
 @NgModule({
   declarations: [
@@ -11,7 +11,11 @@ import { PdfNotFoundComponent } from './pdf-not-found.component';
   ],
   imports: [
     CommonModule,
-    PdfNotFoundRoutingModule
+    PdfNotFoundRoutingModule,
+    NgbModule,
+    NgbCarouselModule,
+    NgbSlide,
+    
   ]
 })
 export class PdfNotFoundModule { }
