@@ -1,23 +1,35 @@
 import { Injectable } from '@angular/core';
 import { HttpClient, HttpParams } from '@angular/common/http';
 import { Observable } from 'rxjs';
+import { API_BASE_URL } from './../api.config';
 
 @Injectable({
   providedIn: 'root'
 })
 export class BookService {
 
+  index = 0;
+
   constructor(
     private http: HttpClient
   ) { }
 
+  public setIndex(index: number) {
+    this.index = index;
+  }
+
+  public getIndex() {
+    return this.index;
+  }
+
+
   public getListBook(): Observable<any> {
-    const url = 'http://54.251.83.205:9091/api/technobanker/v1/backoffice/book/get-books';
+    const url = `${API_BASE_URL}/api/technobanker/v1/backoffice/book/get-books`;
     return this.http.get(url);
   }
 
   public getListBookPartly(): Observable<any> {
-    const url = 'http://54.251.83.205:9091/api/technobanker/v1/backoffice/book/get-books-partly';
+    const url = `${API_BASE_URL}/api/technobanker/v1/backoffice/book/get-books-partly`;
     return this.http.get(url);
   }
 
@@ -28,13 +40,13 @@ export class BookService {
       .set('clientEmail', body.clientEmail)
       .set('clientNumber', body.clientNumber);
 
-    return this.http.post('http://54.251.83.205:9091/api/technobanker/v1/backoffice/book/send-email', params);
+    return this.http.post(`${API_BASE_URL}/api/technobanker/v1/backoffice/book/send-email`, params);
   }
   
   public getDetailBook(body: any): Observable<unknown> {
     const params = new HttpParams()
       .set('bookId', body.bookId);
-    return this.http.get('http://54.251.83.205:9091/api/technobanker/v1/backoffice/book/get-detailed-book', { params: params });
+    return this.http.get(`${API_BASE_URL}/api/technobanker/v1/backoffice/book/get-detailed-book`, { params: params });
   }
 
   saveEmailData(userData: any) {
@@ -44,7 +56,13 @@ export class BookService {
   public previewImage(body: any): Observable<unknown> {
     const params = new HttpParams()
       .set('bookId', body.bookId);
-    return this.http.get('http://54.251.83.205:9091/api/technobanker/v1/backoffice/book/get-detailed-book/book-preview', { params: params });
+    return this.http.get(`${API_BASE_URL}/api/technobanker/v1/backoffice/book/get-detailed-book/book-preview`, { params: params });
+  }
+
+  public updateViewCount(body: any): Observable<any>{
+    const formData = new FormData();
+    formData.append("bookId", body.bookId)
+    return this.http.patch(`${API_BASE_URL}/api/technobanker/v1/backoffice/book/count-views`, formData);
   }
 
 }
