@@ -1,6 +1,7 @@
 import { Injectable } from '@angular/core';
 import { HttpClient, HttpParams } from '@angular/common/http';
 import { Observable } from 'rxjs';
+import { API_BASE_URL } from './../api.config';
 
 @Injectable({
   providedIn: 'root'
@@ -14,7 +15,7 @@ export class HomeServiceService {
     ) { }
 
   public getSlider(): Observable<any> {
-    const url = 'http://54.251.83.205:9091/api/technobanker/v1/backoffice/slider/get-slider-for-frontoffice';
+    const url = `${API_BASE_URL}/api/technobanker/v1/backoffice/slider/get-slider-for-frontoffice`;
     return this.http.get(url);
   }
 
